move MatrixUtilities to ncmath
