+++ include/ncengine/physics/RigidBody.h
@@ -8,7 +8,6 @@
 #include "ncengine/ecs/Transform.h"
 #include "ncengine/physics/Constraints.h"
 #include "ncengine/physics/Shape.h"
-#include "ncengine/utility/MatrixUtilities.h"
 
 namespace nc
 {
