--- include/ncmath/DirectXMathUtility.h
@@ -1,19 +1,0 @@
-/**
- * @file DirectXMathUtility.h
- * @copyright Jaremie Romer and McCallister Romer 2024
- */
-#pragma once
-
-#include "DirectXMath.h"
-
-namespace nc
-{
-inline float GetMaxScaleExtent(DirectX::FXMMATRIX matrix)
-{
-    auto xExtent_v = DirectX::XMVector3Dot(matrix.r[0], matrix.r[0]);
-    auto yExtent_v = DirectX::XMVector3Dot(matrix.r[1], matrix.r[1]);
-    auto zExtent_v = DirectX::XMVector3Dot(matrix.r[2], matrix.r[2]);
-    auto maxExtent_v = DirectX::XMVectorMax(xExtent_v, DirectX::XMVectorMax(yExtent_v, zExtent_v));
-    return sqrt(DirectX::XMVectorGetX(maxExtent_v));
-}
-}
